+++ server/routes/forms.js
@@ -1,3 +1,5 @@
+const csrfCheck = require('./csrf-token-check');
+
 module.exports = class Forms {
     constructor(db) {
         this.db = db;
@@ -13,6 +15,9 @@ module.exports = class Forms {
     }
 
     post(req, res) {
+        if (!csrfCheck(req.headers['csrf-token'], res)) {
+            return;
+        }
         this.db.collection('Forms').insertOne({title: req.body.title, fields: req.body.fields}, (err, result) => {
             if (err) throw err;
             res.sendStatus(200);
@@ -20,6 +25,9 @@ module.exports = class Forms {
     }
 
     delete(req, res) {
+        if (!csrfCheck(req.headers['csrf-token'], res)) {
+            return;
+        }
         this.db.collection('Forms').deleteOne({title: req.body.title}, (err, result) => {
             if (err) throw err;
             res.sendStatus(200);
@@ -27,6 +35,9 @@ module.exports = class Forms {
     }
 
     put(req, res) {
+        if (!csrfCheck(req.headers['csrf-token'], res)) {
+            return;
+        }
         const query         = { title: req.body.title };
         const updatedValues = { fields: req.body.fields, title: req.body.newTitle };
